server/efact/requirements.txt
# Facturación electrónica SUNAT / UBL 2.1 (Python 3.10+)
lxml>=5.0.0
cryptography>=42.0.0
signxml>=3.2.0
requests>=2.31.0
reportlab>=4.0.0
python-dotenv>=1.0.0
